+++ nautilus_trader/adapters/binance/execution.py
@@ -313,7 +313,7 @@ class BinanceSpotExecutionClient(LiveExecutionClient):
                 time_in_force=time_in_force,
                 quantity=str(order.quantity),
                 price=str(order.price),
-                iceberg_qty="0" if order.is_hidden else None,
+                iceberg_qty=str(order.display_qty) if order.display_qty is not None else None,
                 new_client_order_id=order.client_order_id.value,
                 recv_window=5000,
             )
@@ -337,7 +337,7 @@ class BinanceSpotExecutionClient(LiveExecutionClient):
                 quantity=str(order.quantity),
                 price=str(order.price),
                 stop_price=str(order.trigger),
-                iceberg_qty="0" if order.is_hidden else None,
+                iceberg_qty=str(order.display_qty) if order.display_qty is not None else None,
                 new_client_order_id=order.client_order_id.value,
                 recv_window=5000,
             )